update header with long

+++ BadgerApp/BadgerPrefHandler.h
@@ -12,18 +12,18 @@
 
 void badgerSaveUniversalPref(NSString *prefKey, id prefValue);
 void badgerSaveAppPref(NSString *prefApp, NSString *prefKey, id prefValue);
-void badgerSaveUniversalCountPref(int count, NSString *prefKey, id prefValue);
-void badgerSaveAppCountPref(int count, NSString *prefApp, NSString *prefKey, id prefValue);
+void badgerSaveUniversalCountPref(long count, NSString *prefKey, id prefValue);
+void badgerSaveAppCountPref(long count, NSString *prefApp, NSString *prefKey, id prefValue);
 void badgerRemoveUniversalPref(NSString *prefKey);
 void badgerRemoveAppPref(NSString *prefApp, NSString *prefKey);
-void badgerRemoveUniversalCountPref(int count, NSString *prefKey);
-void badgerRemoveAppCountPref(int count, NSString *prefApp, NSString *prefKey);
+void badgerRemoveUniversalCountPref(long count, NSString *prefKey);
+void badgerRemoveAppCountPref(long count, NSString *prefApp, NSString *prefKey);
 void badgerSetUpPrefPlist(void);
 void badgerSetUpPrefPlistAtSpecificLocation(NSString *specifiedDirectory);
 id badgerRetriveUniversalPref(NSString *prefKey);
 id badgerRetriveAppPref(NSString *prefApp, NSString *prefKey);
-id badgerRetriveUniversalCountPref(int count, NSString *prefKey);
-id badgerRetriveAppCountPref(int count, NSString *prefApp, NSString *prefKey);
+id badgerRetriveUniversalCountPref(long count, NSString *prefKey);
+id badgerRetriveAppCountPref(long count, NSString *prefApp, NSString *prefKey);
 NSArray *badgerRetriveConfigsWithUniversalPref(NSString *prefKey);
 NSArray *badgerRetriveConfigsWithAppPref(NSString *prefApp, NSString *prefKey);
 #endif /* BadgerPrefHandler_h */
